+++ spec/integration/sequence_and_naming.rb
@@ -151,6 +151,7 @@ STORAGE_POOL_NAME = 'CPG-SSD-AO'.freeze
 
 # VolumeTemplate
 VOL_TEMP_NAME = 'VolumeTemplate_1'.freeze
+VOL_TEMP_VIRTUAL_NAME = 'VolumeTemplateVirtual_1'.freeze
 VOL_TEMP_NAME_UPDATED = 'VolumeTemplate_1_UPDATED'.freeze
 
 # Volume
@@ -158,7 +159,11 @@ VOLUME_NAME = 'Volume_1'.freeze
 VOLUME2_NAME = 'Volume_2'.freeze
 VOLUME3_NAME = 'Volume_3'.freeze
 VOLUME4_NAME = 'Volume_4'.freeze
+VOLUME5_NAME = 'Volume_5'.freeze
+VOLUME6_NAME = 'Volume_6'.freeze
 VOLUME7_NAME = 'Volume_7'.freeze
+VOLUME_VIRTUAL_NAME = 'Volume_Virtual_1'.freeze
+VOLUME_VIRTUAL2_NAME = 'Volume_Virtual_2'.freeze
 VOL_SNAPSHOT_NAME = 'snapshot_qa'.freeze
 VOL_SNAPSHOT2_NAME = 'snapshot_qa_2'.freeze
 
@@ -190,6 +195,9 @@ SERVER_PROFILE_WITH_OSDP_NAME = 'ServerProfile_OSDP'.freeze
 
 # Server Profile Template
 SERVER_PROFILE_TEMPLATE_NAME = 'ServerProfileTemplate_1'.freeze
+SERVER_PROFILE_TEMPLATE2_NAME = 'ServerProfileTemplate_2'.freeze
+SERVER_PROFILE_TEMPLATE3_NAME = 'ServerProfileTemplate_3'.freeze
+SERVER_PROFILE_TEMPLATE4_NAME = 'ServerProfileTemplate_4'.freeze
 SERVER_PROFILE_TEMPLATE_NAME_UPDATED = 'ServerProfileTemplate_1_UPDATED'.freeze
 CONNECTION_NAME = 'EthernetConnection'.freeze
 CONNECTION2_NAME = 'FibreChannelConnection'.freeze